Include storage id's in data_distribution_stats
