+++ fdbclient/SpecialKeySpace.actor.cpp
@@ -348,28 +348,3 @@ Future<Standalone<RangeResultRef>> ConflictingKeysImpl::getRange(ReadYourWritesT
 	}
 	return result;
 }
-
-ACTOR Future<Standalone<RangeResultRef>> ddMetricsGetRangeActor(ReadYourWritesTransaction* ryw, KeyRangeRef kr) {
-	auto keys = kr.removePrefix(ddStatsRange.begin);
-	Standalone<VectorRef<DDMetricsRef>> resultWithoutPrefix =
-	    wait(waitDataDistributionMetricsList(ryw->getDatabase(), keys, CLIENT_KNOBS->STORAGE_METRICS_SHARD_LIMIT));
-	Standalone<RangeResultRef> result;
-	for (const auto& ddMetricsRef : resultWithoutPrefix) {
-		// each begin key is the previous end key, thus we only encode the begin key in the result
-		KeyRef beginKey = ddMetricsRef.beginKey.withPrefix(ddStatsRange.begin, result.arena());
-		// Use json string encoded in utf-8 to encode the values, easy for adding more fields in the future
-		json_spirit::mObject statsObj;
-		statsObj["ShardBytes"] = ddMetricsRef.shardBytes;
-		std::string statsString =
-		    json_spirit::write_string(json_spirit::mValue(statsObj), json_spirit::Output_options::raw_utf8);
-		ValueRef bytes(result.arena(), statsString);
-		result.push_back(result.arena(), KeyValueRef(beginKey, bytes));
-	}
-	return result;
-}
-
-DDStatsRangeImpl::DDStatsRangeImpl(KeyRangeRef kr) : SpecialKeyRangeAsyncImpl(kr) {}
-
-Future<Standalone<RangeResultRef>> DDStatsRangeImpl::getRange(ReadYourWritesTransaction* ryw, KeyRangeRef kr) const {
-	return ddMetricsGetRangeActor(ryw, kr);
-}
